feat: localized NoDataView
